jta: Fix template-dev target DISTRIB env variable
The current front-end Jenkins configuration for the template-dev
board currently uses DISTR instead of DISTRIB as an environment
variable to specify the distribution overlay file to use.

This caused an error when processing tests for any targets copied
from that board, about the missing DISTRIB environment variable.
From the console log:

  +++ abort_job 'DISTRIB is not defined'
  +++ set +x

  *** ABORTED ***

  JTA error reason: DISTRIB is not defined

Also, this patch adds a missing newline at the end of config.xml