Add a script to maintain the symlink for config.xml into,/userdata/conf
This adds a feature to maintain the symlink from /var/lib/jenkins/config.xml
to /userdata/conf/config.xml.  Jenkins removes this symlink when it
edits the file (which happens when a board is added or removed).

This change adds the inotify-tools package to the container build,
and adds a shell script which uses that in order to monitor the
directory for changes, and repair the link if needed.  A small
change is made to the jta-start-cmd.sh to start the script
(which runs forever).

The script is currently a bit noisy, but I could add a 'quiet'
option if desired, so that it runs silently.