Change tools.sh to read individual $PLATFORM-tools.sh scripts
This allows someone to add a toolchain without editing tools.sh
The will allow toolchains to be packaged for fuego more easily.

The new method ot add a toolchain is to install it under
/userdata/toolchains, then add a file called <platform-name>-tools.sh
in /userdata/toolchains, which sets the variables needed for the
toolchain.  For Yocto project-based toolchains, this involves
calling the environment-setup script for the SDK that is built with YP.

I migrated the lager and poky-qemu-arm PLATFORMS as examples.