Fix problem with docker build fetching jenkins-cli.jar too quickly
updates.sh is called during the docker build to modify some portions
of jenkins (to customize it for fuego).  This script downloads
jenkins-cli.jar, in order to use that to install some modified plugins.
However, updates.sh is called by a dockerfile command immediately
after a jenkins restart.

On some systems, when the container is being built it takes a
while before jenkins becomes available.  Add a delay to give time
for jenkins to come up, in sync_fetch_jenkins_cli().

Also, this routine was looping forever.  Add a retry, to avoid
an infinite loop.