Add python-xmltodict for some easier xml handling
This is used by the experimental functional parser.py, and could
be used to simplify the create_xml_<test_name>.py files (if we
keep those around.  We should also use this in ftc to simplify
handling of the build.xml file, and possibly other Jenkis xml
files in the future.

Signed-off-by: Tim Bird <tim.bird@sony.com>