toolchains: Use -E to access pre-processor for Debian crosscompiler
The Debian ARM cross-compiler that can be installed with
install_armhf_toolchain.sh does not include arm-linux-gnueabihf-cpp,
but that's what we were telling Fuego to use.
Multiple tests that used qemu-armv7hf-tools.sh were failing in
their configure step, with a message about "C preprocessor
failing sanity check".

Change the definitions for CPP and CXXCPP to add the '-E' compile
flag, which invokes the C and C++ preprocessors, respectively,
to fix this problem.

Signed-off-by: Tim Bird <tim.bird@sony.com>