Version 1.1.0 release - "Blaze"
This is the "Blaze" release of Fuego.  This includes a major
overhaul of the integration between the Fuego core system and
the Jenkins continuous integration tool.  A lot of functionality
has been added to the new 'ftc' tool.  See the CHANGELOG for details
for this release.

This commit updates meta-files for this release, and declares the
release and release name.

Signed-off-by: Tim Bird <tim.bird@sony.com>