Install model file to support JENKINS_URL out of the box
Install a jenkins.model.JenkinsLocationConfiguration.xml file
into the container, so that Jenkins can provide JENKINS_URL,
BUILD_URL and JOB_URL immediately after installation.
Otherwise, the user has to go to Manage Jenkins, Configure Server,
and click "Save" to create this file.

This will allow us to not rely on a hardcoded URL in the Fuego
scripts.