Add volume mount of /boot to docker container
This is useful for Functional.fuego_dependencies to get
the kernel configuration for the kernel running on the host.
This is mounted read-only in the configuration, to minimize
the danger of some security breach of network services inside
the container from endangering the configuration of the machine.

However, it does represent a potential information leak.
If someone is not using the 'docker' test board, it is probably
better not to include this.