flot: add support for HTML table visualization
This patch produces HTML tables per test_case. It is just
an initial proof of concept so that AGL can work on improving
it until it looks like what they have.

The plotting code in flot may not currently be updated to the
latest parser output format (results.json). But we need to
decide if we really want the plotting, or if we rather have
a table with data even for benchmark tests.

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>