Dockerfile: add libraries needed for graphical tests
Add packages for libraries needed for graphical tests
blobsalad (which needs libSDL and libcairo) and
x11perf (which needs libxmuu and libxmu).

This should also be followed up with pre_check tests
in the respective tests, to verify if the required headers
and libraries are available in the SDK (and the libraries
are available on the target).

Signed-off-by: Tim Bird