vuls: add script for installing go on docker
Vuls requires a version of "go" that is newer than the
one provided by Fuego docker's debian. We first prepared
a patch to the Fuego's Dockerfile, but we noticed
that some proxies block URLs under "storage.googleapis.com".
To avoid users having a bad first-time experience when
installing Fuego we decided to provide it as an script instead.

As an alternative we can provide the Dockerfile patch
and ask those users with proxies blocking such URLs to
comment those lines and install go manually.

Reviewed-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>