Install crosstoolchains from debian unstable instead of emdebian
Emdebian has reportedly ceased to publish updates to its packages since
the beginning of 2015 causing some of them to be outdated and to
conflict with other packages in Debian Jessie.

The arm crosstoolchains that we were fetching from emdebian are
currently conflicting with the newest packages from debian/security,
what prevents the build to succeed.

On this patch, we remove emdebian sources and add debian unstable to apt
sources in order to install those packages from debian official
repositories.

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>