Dockerfile: sdk dependencies for building a x86_64 kernel
Note, I have added the dependencies to the kernel_build test's
test.yaml. In the future, it would be nice to install them
directly from ftc. At the moment that is not possible so I add
them here, but I'm also adding is_on_sdk checks to the test
so even if this patch is not accepted the user should be
able to understand the error log.

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>