docker: load plugins using hardcoded versions
Use install-plugins.sh (copied from github jenkinsci/docker project)
to load the plugins for Jenkins using hardcoded versions.

This resolves issues with using the update center with the latest
versions of packages, which yields lots of dependency errors and
makes the description-setter plugin not work (among others).

For this release, use the same Jenkins war version and plugin
versions as the 1.2.1 release (from November, 2017).

Signed-off-by: Tim Bird <tim.bird@sony.com>