Set DEBIAN_FRONTEND=noninteractive globally during build
Use ARG DEBIAN_FRONTEND=noninteractive to set apt-get user interface as
non-interactive, preventing the warning messages "unable to initialize
frontend: Dialog" to show up during build.

By using ARG we make sure that the value of DEBIAN_FRONTEND is set only
during build, and not on the resulting image, or on builds of images
that are based on this one.

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>