board: add custom criteria.json file for beaglebone
By default, the Beaglebone comes with a non-realtime kernel.
Add a board-specific criteria.json file, with max_latency
set very high (11 milliseconds), so this test passes.

If a realtime enabled kernel is used on this board, the
criteria should be adjusted for match a lower expected value
for max_latency.

Signed-off-by: Tim Bird <tim.bird@sony.com>