boards: add ARCH_TIMER variables to some boards
ARCH_TIMER_NAME and ARCH_TIMER_ID are required for the
Functional.arch_timer test.  See that test for details for
what should be put in these variables.

Also add these to the template-dev board file, so users
will know what they need to customize if they use that test.

Signed-off-by: Tim Bird <tim.bird@sony.com>