Add 'localhost' board, to use host as DUT
The 'docker' board is for testing tests inside the docker
container, and it is not running the same distribution as
the host.  The 'localhost' board is used to run Fuego tests
on the host itself.

The current configuration requires a user account 'fuego',
with a home directory of '/home/fuego'.  However, this
can easily be customized by the Fuego user.

Signed-off-by: Tim Bird <tim.bird@sony.com>