tools: make toolchain setup less verbose
'env -u' has the side-effect of printing the environment.
This is counter to what the man page and help say for this
command.  Get rid of output when unsetting the PYTHONHOME
directory.  If we want to see the environment, we'll add
something to the core for this.  Otherwise, toolchain
setup should be silent.

Signed-off-by: Tim Bird <tim.bird@sony.com>