Dockerfile: avoid error adding group
Only add the Jenkins group if it's not already there.
Fuego tries to build the docker image so that the user
'jenkins' inside the image has the same uid and gid as
the user who built the container.  This makes operations
on the volume-mounted directories easier.

However, on some systems the host user's gid may already
be present inside the image (particularly if they are using
Debian in a default configuration).  Don't try to create
the group inside the container if it already exists.
Just use it as is.

Separate the lines so that the useradd is separate
from the groupadd.

Reported-by: Georg Piewald <gpi@mailbox.org>
Signed-off-by: Tim Bird <tim.bird@sony.com>