docs: add a REAME for building the docs
Our in-source docs are very confusing.  I have started the
conversion to .rst (Sphinx), but the Makefile contains
instructions for building both the old (fuego-docs.pdf)
and new (.rst-based) docs.

Probably, the docs should be moved to fuego-core anyway.
And the old docs are very out-of-date.  But for now, at
least let people (including my future self) know how to
build each set of docs.

Signed-off-by: Tim Bird <tim.bird@sony.com>