install: simplify the installation process
We have confusing instructions for building the docker
image and container, and starting it.  Simplify things
by add --help to install.sh, converting it to also
do the container build, and adding a start.sh for
starting the container.

This is simpler and more uniform than the old method of
running install.sh and then the stuff in fuego-host-scripts.

Signed-off-by: Tim Bird <tim.bird@sony.com>