container: add get-pip.py script, to avoid pip3 installation woes
Our current Fuego container setup (in Dockerfile) results in
dependency errors for the python3-requests module:
/usr/local/lib/python3.4/dist-packages/requests/__init__.py:91: \
  RequestsDependencyWarning: urllib3 (1.9.1) or chardet (3.0.4) doesn't
    match a supported version!
  RequestsDependencyWarning)

This shows up when trying to run the fuego_release_test, or when
running 'pip3' manually inside the container.

While not a full solution, add the get-pip.py script to manually
install pip3 (and requests and urllib3 and chardet).  I'll hold
off modifying Dockerfile to use this, until after the 1.4 release
(as we may be switching to 'stretch' soon anyway, which might resolve
this issue).  But this issue needs to get fixed soon after the 1.4
release.

Signed-off-by: Tim Bird <tim.bird@sony.com>