build: add NO_CACHE (in comment) for testing build scenarios
You can use --no-cache to force "docker build" to not use the
docker image cache.  This slows the build down, but is useful
when trying to reproduce certain build failures.

I *could* make this a command line argument, but that's too
much complexity for how often I use this in debugging.
For now, just add the option, so it's easy to turn on or off
by uncommenting the line with the NO_CACHE shell variable.

Signed-off-by: Tim Bird <tim.bird@sony.com>