docker-container: allow user to override JENKINS_PORT
Install.sh already supports specifying the port for Jenkins.
That is useful when you want to install Fuego on a machine
that is already using 8080.

Another use case is to install multiple instances of Fuego
on the same machine. Creating one image per jenkins port
is quite inefficient. Instead, we want to create a container
for each port and override the Jenkins port.

This patch allows overriding the environment variable
JENKINS_PORT, used by ftc or common.sh for example. However,
it requires an extra step (modify jenkins configuration
file) which needs to be done manually (see the script).

It would be nice to automatize that step with the parameter
--entrypoint but I haven't been succesful yet.

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>