Revert "jenkins: disable certain node monitors for docker nodes"
It appears that the notifications and node offlining were due
to a problem in my local setup.  Remove this "fix", and leave
these notification monitors enabled.  I'll revisit this and
re-install the fixup program if other people see this problem
(if it turns out to not be unique to my setup.)

This reverts commit 548b300b419a1fd5a8797482259b7df3a9b844ca.

Signed-off-by: Tim Bird <tim.bird@sony.com>