toolchains: remove ORIG_PATH from non-yocto setup scripts
ORIG_PATH is used by Fuego to get access to the system
python interpreter.  Yocto environment setup scripts alter
the PATH, adding an element that puts a board-specific
python interpreter on the PATH.  However, for Fuego operations
we want to use the system interpreter and not the toolchain-specific
interpreter.

Remove setting ORIG_PATH in toolchain setup scripts which do not
mess with the system PATH, as this is not needed.

Signed-off-by: Tim Bird <tim.bird@sony.com>