serio: Fix serio @ handling and live streaming
Apply 2 patches to serio to fix it's handling of '@' embedded in a
serial device name, and to support live streaming of data during
an operation (that is, to avoid buffering the data for the duration
of command operation).  See the patches for individual descriptions.

These changes are applied as patches against the upstream serio
repo.  The changes were submitted, but not accepted, and I didn't
have time before this release to finish upstreaming these patches
(although Frank was receptive to having both of these items fixed).

So, adjust Dockerfile to apply the serio patches as part of Fuego
installation.  This is a bit fragile, as upstream might break the
patches, but I'll deal with that when it happens.

Signed-off-by: Tim Bird <tim.bird@sony.com>